qutemail-backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-decouple==3.8
cryptography==41.0.7
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0
dj-database-url==2.1.0
mail-parser==3.15.0
